move tests out to tests directory
not everything needs to be in one file
testing public interfaces, so no need to be within lib
